GAS: wire final hardcoded movement/chat/bodypart/hacking values
- chatDisplayTicks (GameEngineDef, was showchat_i=255): chat overlay duration
- walkAcceleration (PlayerDef, was xv+=3): walk/run acceleration per tick
- disguisedDecelSpeed / disguisedDecelSpeedSecret (PlayerDef, was 4/2): xv snap on deceleration while disguised
- ladderSpeedReduction (PlayerDef, was -4): speed handicap on ladders; ladder state now uses GAS runSpeed variants
- hackingPowerupBonus (PlayerDef, was 1.00): hacking speed multiplier from powerup
- bodyPartSpawnYOffset / bodyPartLaunchYV (WorldDef, was 50/20): body part death physics

Change world.h showchat_i from Uint8→int to allow values >255 if tuned.
Update gameengine.json, world.json, player.json with matching defaults.

+++ clients/silencer/src/world/world.h
@@ -93,7 +93,7 @@ public:
 	bool replaying;
 	Uint8 quitstate;
 	std::deque<std::string> chatlines;
-	Uint8 showchat_i;
+	int showchat_i;
 	std::vector<BuyableItem *> buyableitems;
 	std::deque<char *> statusmessages;
 	static const int maxstatusmessages = 4;
